Move scripts to layout
