+++ src/app/(website)/layout.tsx
@@ -1,6 +1,7 @@
 import type { Metadata } from "next";
 import { Providers } from "../(website)/providers";
 import { DataLayer } from "@/components/DataLayer";
+import Script from "next/script";
 
 export const metadata: Metadata = {
   title: "Kokua",
@@ -25,6 +26,33 @@ export default function RootLayout({
           <Providers>
             <DataLayer>{children}</DataLayer>
           </Providers>
+
+          <Script id="show-banner">
+            {`window.plausible = window.plausible || function() { (window.plausible.q = window.plausible.q || []).push(arguments) }`}
+          </Script>
+
+          <Script id="show-banner">
+            {`
+            
+            function prepareUrl(params) {
+              const url = new URL(location.href)
+              const queryParams = new URLSearchParams(location.search)
+              let customUrl = url.protocol + "//" + url.hostname + url.pathname.replace(/\/$/, '')
+              
+              if(params){
+                for (const paramName of params) {
+                  const paramValue = queryParams.get(paramName)
+                  if (paramValue) customUrl = customUrl + '/' + paramValue
+                }
+              }
+
+              return customUrl
+            }
+            plausible('pageview', { u: prepareUrl(["tag", "query"]) + window.location.search })
+          
+
+            `}
+          </Script>
         </body>
       </html>
     </>
