+++ src/app/(website)/providers.tsx
@@ -6,7 +6,6 @@ import { ChakraProvider } from "@chakra-ui/react";
 import { DM_Sans } from "next/font/google";
 import PlausibleProvider from "next-plausible";
 import { AppProgressBar as ProgressBar } from "next-nprogress-bar";
-import Script from "next/script";
 
 const dm_sans = DM_Sans({ subsets: ["latin"] });
 
@@ -28,33 +27,6 @@ export function Providers({ children }: { children: React.ReactNode }) {
             src: "https://plausible.io/js/script.manual.js",
           }}
         >
-          <Script id="show-banner">
-            {`window.plausible = window.plausible || function() { (window.plausible.q = window.plausible.q || []).push(arguments) }`}
-          </Script>
-
-          <Script id="show-banner">
-            {`
-            
-            function prepareUrl(params) {
-              const url = new URL(location.href)
-              const queryParams = new URLSearchParams(location.search)
-              let customUrl = url.protocol + "//" + url.hostname + url.pathname.replace(/\/$/, '')
-              
-              if(params){
-                for (const paramName of params) {
-                  const paramValue = queryParams.get(paramName)
-                  if (paramValue) customUrl = customUrl + '/' + paramValue
-                }
-              }
-
-              return customUrl
-            }
-            plausible('pageview', { u: prepareUrl(["tag", "query"]) + window.location.search })
-          
-
-            `}
-          </Script>
-
           <ChakraProvider theme={theme}>
             {children}
